Support building a single executable with full testsuite
This adds the "makealltests" script which generates a .c file including
all test files, plus support in "Makefile" to build and run it. Now it
is possible to just do:

	make run-all-tests

And that will create a single binary, which is optimal when one wants to
run the entire test suite at once. The option to build independent suites
is kept, though.