wopt: Fix W_OPT_BOOL to set value instead of toggling
Toggling values might be confusing, so it is better to just store a truth
value when a flag is encountered. Alternatively, we could add an
W_OPT_BOOL_NO to store W_NO, and/or W_OPT_BOOL_TOGGLE if the old behavior
is desired. Another option would be allowing "--no-<option>" to set W_NO,
but that would require further changes in the option lookup code.