Implement simple, no-frills socket using w_io_t
For the moment only TCPv4 and Unix domain sockets are supported, due to
the stream-based nature of operations supported by w_io_t objects. The
API is simple enough that:

- Making a (simple) server does only imply calling w_io_socket_open
  and w_io_socket_serve with an user-supplied request handling function.

- Making a (simple) client does only imply calling w_io_socket_open
  and w_io_socket_connect.

Once the sockets are created, all w_io_* and w_io_socket_* functions may
be called on them. The nice thing about the API is that, once the w_io_t
object is created, it does not matter whether one is handling a file, a
TCP socket or an Unix domain socket.