Add support for a "putback" character un I/O streams
Sometimes it is desirable to make some backtracking over an input stream.
This patch adds support for a single "putback" character: if a putback
character has been defined, then that will be the next returned by reading
the stream. This satisfies the same needs as the ungetc() function does for
FILE* in the C standard library.