Fix compilation warning on some GLIB versions.
The example code for sockets fails with error:

error: ignoring return value of 'write', declared with
       attribute warn_unused_result

It could be a problem to use write (whis is declared to warn on unused
result) without catch the return value, while compiling with -Werror.

This fix sets the ret variable to this return. If write do not fails, ret
value will be kept.