Improvements to work with dietlibc
This patch adds improvemnts to work fine with dietlibc, including SUN_LEN
macro, and initialization on some unitialized variables.

With this changes libwheel compiles fine (and run) with dietlibc version of
Debian and Ubuntu. These versions, AFAIK also has math patches, so to be
dietlibc total compatible probably a definition of math INFINTY and NAN
would be required.

Signed-off-by: Adrian Perez <aperez@igalia.com>