wobj: Allow using w_obj_{ref,unref} in static objects
When all the bits of an object's reference counter are set (-1 in two's
complement), then the w_obj_ref() and w_obj_unref() will not update the
reference counter. When unreferencing, if a destruction function exists, it
will be called, and then set to NULL (to avoid further calls).

This patch also introduces a new W_OBJ_STATIC() macro to aid in creation of
static objects.

Support for static objects is useful for creating objects in the stack, or
statically preinitialized.