wio-unix: New w_io_unix_{open,init} and rename w_io_unix_{open,init}_fd
Existing functions working directly with file descriptors have now the *_fd suffix,
and old names are used for functions that perform the call to open(2) themselves.
The most common use case will be opening w_io_unix_t's by path and mode, so it
makes sense to be the default.