wbuf: Avoid compiler warning in w_buf_length
As usually w_buf_t structures are allocated on the stack, checking
for the address of it will always succeed. Some compilers will
give a warning on this. Anyway, for non-stack allocated structs
or for those cases where the compiler doesn't know that the address
is constant, it's not a big deal to skip the check, as a quick
inspection with a debugger will reveal the issue.