wlist: Optionally support refcounting elements
When lists are created with "w_list_new (W_YES)", it will be assumed that
all contained elements are objects (derived from w_obj_t), and when added
(removed) to (from) the list, the item's reference conter will be incressed
using w_obj_ref (w_obj_unref). This is a nice shorthand to maintain lists
of objects.