wdict: Optionallly support updating refcounts
Adds a new argument to w_dict_new, which can be used to enable automatic
updating of object reference counts when added/removed to a dictionary.
This works in the same way as for the list container, and brings
dictionaries on-par with it. It is assumed that all values added when
the option is enabled derive from w_obj_t.