wio-buf: Fix bug in buffer sharing
Now the w_io_buf_t holds a pointer to the buffer being used:

- Points to the internal w_buf_t when NULL is passed to w_io_buf_{open,init}
- Or it may point to any other buffer.

Previously, the contents of the passed buffer were copied inside the
internal one. This was okayish (bot not completely correct) when the
passed buffer already conained some data, and it failed horribly for
empty buffers. Now, having an extra indirection level, buffer sharing
is done properly, and without the pitfall of not having buffer sizes
updated when the buffer was manipulated manually instead of using
w_io_{read,write,...}