wvariant: Do not use fixed type enum values
Values in the w_variant_type_t enum are now not manually defined. The
initial idea was to reuse those definitions to ease the implementation
of (de)serialization to tnetstrings, but in the end that would be exposing
an implementation detail, so this patch removes the manual definitions and
lets the compiler chose the enumeration values.