wtnetstr: Implement writing tnetstrings to w_io_t
Primitive values are written directly to output streams, which is good
for saving memory as no intermediate buffers will be allocated. This applies
to strings, booleans, buffers and nulls. The rest of the values still need
to be formatted in a buffer before writing so the final tnetstring size is
known.