wtnetstr: Implement reading tnetstrings from w_io_t
Provides convenience functions to deserialize tnetstrings from w_io_t
streams. The main workhorse is in w_tnetstring_read_to_buffer(), which:

- Reads the size of the tnetstring payload, one character at a time, until
	five digits have been consumed or a colon is found.
- Once the size of the payload is known, a bulk read of the size of the
	payload (plus one byte for the terminating type tag) is done.

The rationale is to make it so one message is read at a time, and blocking
is done only until the full tnetstring is read. This is interested to use
tnetstrings as basis for messages between different processes, or over TCP
sockets.