wio: Fix w_io_fscan so it actually works
This fixes flaws in function w_io_fscan and improves it so it is actually
useful and not just a mere convenience to invoke the other w_io_fscan_*
functions in a sequence. More precisely:

* The format pattern can now contain an arbitrary sequence of to-match
  characters and format specifiers, i.e. the following is valid:

       w_io_fscan (io, "pi=$F\ne=$F", &pi_val, &e_val)

  It will first match the literal sequence of characters "pi=", then read
  a floating point number into "pi_val", then match the "\ne=" character
  sequence, and finally read a floating point number into "e_val".

* On error, the lookahead character is put back in the stream with
  w_io_putback.

* The number of values correctly converted is returned. The above
  example would return "2" if both values are successfully read.

The two later fixes also enable to continue using the stream on conversion
failures and still not lose bytes from the input stream.