wobj: Allow passing NULL to w_obj_(ref|unref)
Some parts of the code in libwheel (especially some of the code merged
lately for the w_variant_t support) assumes that it is safe to pass NULL
to w_obj_ref and w_obj_unref. Actually allowing it makes sense and makes
it possible to write shorter code that updates the reference counters.
This patch enables passing NULL to the functions.