wobj: Allow manually marking objects as static
This prevents the final w_free() of instances when unrefing/destroying them,
and in turns allows to have object structs as part of other object (not as
a pointer, but the struct directly in the other object). Note that this does
not play well with w_obj_new_with_priv_sized().