wmeta: Introduce new definitions to describe types
This introduces the w_meta_type_t (and related enum values), w_meta_item_t
and w_meta_desc_t, to describe compound data types. The idea is to allow
declaring meta information of types in a rather flexible way. Example:

  typedef struct {
    char *name;
  } Job;

  typedef struct {
    char   *name;
    char   *lastname;
    uint8_t age;
    Job     job;
  } Person;

  w_meta_t person_meta = {
  	W_META ("Person"),
  	W_META_STR (Person, name),
  	W_META_STR (Person, lastname),
  	W_META_U8  (Person, age),
  	W_META_STR (Person, job.name),
  	W_META_END
  };

Note that metadata can be nested, and the above could be written like this
instead:

  w_meta_t job_meta = {
  	W_META ("Job"),
  	W_META_STR (Job, name),
  	W_META_END
  };

  w_meta_t person_meta = {
  	W_META ("Person"),
  	/* all other fields */
  	W_META_REG (Person, job, job_meta),
  	W_META_END
  };

So in the end it is up to the programmer what is their preferred way of
defining the fields. The first approach may be preferred for simpler cases,
and the second for complex nested structures.

The long term goal is to provide conversion functions that can (de)serialize
structures automatically (e.g. from/to tnetstrings). Note that nothing
prevents from describing objects as created with W_OBJ(), as long as all
fields of interest are properly described.