wevent: Put idle events in its own list
Idle events are not passed down to the back-end (epoll/kqueue) and they are
handled by the wevent subsystem in user space. Provided that we do not get
events from the kernel for them, it is better to have them in a separate
list, to avoid having to scan the list with all other events to run them.