wtask: Rename w_task_spawn() to w_task_prepare()
Using "spawn" gives the idea that the task is started right away, but that
does not happen before w_task_run_scheduler() is called. Using "create" in the
name would be better, but it implies that something is created that client
code will have to free/destroy later. In the case of tasks, resources are
allocated but this is transparent to the user, because they will be freed
automatically, so using "prepare" in the name seems to be the best fit.