wparse: Do not keep an extra ref to the input stream
The input stream is used only during the execution of w_parse_run(), so as
long as the caller keeps the object alive, w_parse_run() itself does not need
to take care of it. To make sure client code does not pick the stream from the
w_parse_t structure, the "input" field is NULL-ified upon return.